fixes the bugs
