+++ src/app/globals.css
@@ -102,55 +102,3 @@ a {
   text-decoration: none;
   -webkit-tap-highlight-color: transparent;
 }
-
-.faqs {
-  width: 100%;
-  max-width: 768px;
-  margin: 0 auto;
-  padding: 15px;
-  text-align: left;
-}
-
-.faqs .faq {
-  margin: 15px;
-  padding: 15px;
-  background: #F2F2F2;
-}
-
-.faqs .faq .faq-question {
-  position: relative;
-  font-size: 20px;
-  padding-right: 80px;
-  transition: all 0.4s ease;
-}
-
-.faqs .faq .faq-question::after {
-  content: "+";
-  position: absolute;
-  top: 50%;
-  right: 0px;
-  transform: translateY(-50%);
-  width: 30px;
-  height: 30px;
-  transition: all 0.2s ease;
-}
-
-.faqs .faq .faq-answer {
-  opacity: 0;
-  max-height: 0;
-  overflow: hidden;
-  transition: all 0.2s ease;
-}
-
-.faqs .faq.open .faq-question {
-  margin-bottom: 15px;
-}
-
-.faqs .faq.open .faq-question::after {
-  content: "-";
-}
-
-.faqs .faq.open .faq-answer {
-  max-height: 1000px;
-  opacity: 1;
-}
